Move Event related entites to separate package
