+++ ippon/models/tournament.py
@@ -4,7 +4,7 @@ from math import floor
 from django.db import models
 
 import ippon.models.player as plm
-import ippon.event_models as em
+import ippon.models.event as em
 
 
 class TournamentAdmin(models.Model):
